Fix non-c90 'z' printf modifier for size_t

+++ samples/multiloop_test.c
@@ -14,13 +14,13 @@
 
 void test_callback(const helium_connection_t *conn, uint64_t sender_mac, char * const message, size_t count)
 {
-  helium_dbg("1 Function-pointer callback got %s %zd\n", message, count);
+  helium_dbg("1 Function-pointer callback got %s %lu\n", message, (unsigned long)count);
   helium_dbg("1 Mac address is %" PRIu64 "\n", sender_mac);
 }
 
 void test_callback2(const helium_connection_t *conn, uint64_t sender_mac, char * const message, size_t count)
 {
-  helium_dbg("2 Function-pointer callback got %s %zd\n", message, count);
+  helium_dbg("2 Function-pointer callback got %s %lu\n", message, (unsigned long)count);
   helium_dbg("2 Mac address is %" PRIu64 "\n", sender_mac);
 }
